move assume role helper to parent module. and refacor too many copy pest
